requirements.txt
Flask~=3.1.1
python-dotenv~=1.1.1
flask-cors~=6.0.1